Fix for AJAX tab changes
This is done through a setInterval, which is pretty hacky and I'd prefer
a better way to do it. However, the tabs change through
history.pushState calls, which don't seem to trigger any kind of event.
There are some hacks that involve monkey-patching the pushState function
but I'm not sure if that's any better.